moved fireSearchEvent into custom hook
